Add Scripts, Assets, and ProjectSettings
The Assets (and Scripts), and ProjectSettings folders have been
added back to the repository. This has been done in order to make
sure that we are not adding device-specific code to the repository.

The .gitignore file should only be pulling these files, while, the rest
in the repository should remain untouched.