Add modified .gitignore files and VS sln.
The visual studio solution file should also be tracked with the repo.
This is because the .sln file allows developers to open up the project
file outside of Unity.

With this commit, a fresh clone will result in a useable repo without
the per-system files that don't need to be tracked with source control.